LC-coin-change-ii [Runtime: N/A; Memory: N/A]

+++ dynamic-programming/coin-change-ii.cpp
@@ -4,7 +4,7 @@ public:
         int n = coins.size();
         if(n == 0) return 0;
 
-        vector<unsigned int> prev(amount + 1, 0);
+        vector<unsigned long long int> prev(amount + 1, 0);
         
         for(int a = 0; a <= amount; a++) {
             if(a % coins[0] == 0) {
@@ -13,10 +13,10 @@ public:
         }
 
         for(int i = 1; i < n; i++) {
-            vector<unsigned int> curr(amount + 1, 0);
+            vector<unsigned long long int> curr(amount + 1, 0);
             for(int j = 0; j <= amount; j++) {
-                int notPick = prev[j];
-                int pick = 0;
+                unsigned long long int notPick = prev[j];
+                unsigned long long int pick = 0;
                 if(coins[i] <= j) {
                     pick = curr[j - coins[i]];
                 }
